Updates to the Finnish translation (#618)
* Updates to the Finnish translation

A lot of missing additions and some tweaking. Additionally, a few %1 variables that are used in place of FDM (or Free Download Manager) were replaced with actual text. In Finnish we MUST be able to inflect names too, even foreign ones, to be able to translate sentences correctly. Check out the following recourse for more information: https://uusikielemme.fi/finnish-grammar/finnish-cases/interesting-inflection/inflection-of-foreign-names-in-finnish

This does not change their meaning or make them obscure - instead it has the opposing effect and makes the text more fluent and natural to read.

However, how this affects (if at all) building the final language files is a separate topic and if this poses a problem, please notify me and I will try to somehow make the sentences work with these variables.

* Updates to the Finnish translation

Discussed fixes to the Finnish translation.

* Updates for Finnish

Recent additions and some older missing strings have been added. Plus a few fixes and changes.

* Updates to the Finnish translation

Some minor fixes and changes/polishing.